backend/requirements.txt
Flask==3.0.0
werkzeug==3.0.1
Flask-CORS==4.0.0
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0

